Develop to main check (#68)
* Remove wpa_supplicant copy repository

* wpa_supplicant added as submodule

* wifi detection enhancement for mesh-11s and mesh-ibss scripts

* openwrt as submodule

* use https instead of git

* Revert "use https instead of git"
This reverts commit 34d240ff7b5ea766e7b4cdd6eb8ca0b81f729af3.

* openwrt submodule updated to the latest commit

* User and mesh configuration is now taken from mesh_com.conf yaml file.

* wpa/openwrt submodules updated to the latest content (#58)

* mesh_com: mesh_com debian removal script added (#59)

* mesh_com: Fix STA connection issue post mesh configuration

IBSS and sta interface are not supported concurrently
with most of single radio(phyx). In order to allow sta
connection post mesh service is running, stop mesh(IBSS)
service on the selected interface if already running.


* mesh_com:utils: Add phy parameter monitoring utility infra

Add utility infra to capture or periodically monitor/log various
mac/phy parameter that includes radio rssi, CSI, BW, RF temperature,
PER, peer stats etc.

This change adds RSSI monitoring support on a given interface with
configurable periodicity.

uses:monitor_phy_param.py -r period -i interface
log file: /var/logs/rssi2021-06-29 10:53:34.833729.txt


* mesh_com:utils: use argparse for cmdline argument parsing

Use standard argparse library for command line argument parsing.


* mesh_com:utils: fix a typo

Fixes a typo.


* mesh_com:utils: log time stamp with each RSSI sample

Log time stamp along with rssi captured.
Fix logfile name formatting('%m_%d_%Y_%H_%M_%S').


* mesh_com:utils: Add support for MAC OUI extraction

Extract mac oui from mac address to use this as
trusted element.


* mesh_com:utils: Add support for CSI capture

Add CSI capture support for brcm43455 chipset.
As of now only Broadcomm/Cypress WiFi chipset
found in RPI3B+ is supported.


* Search for AP and mesh interfaces using yaml conf

* Better protection for opening yaml conf. Make server dfltrt configurable.

* Fix because I turned off a bunch of things in the conf and then merged by accident

* When you verify  the list(interface), since it's a filter, you lost it

* mesh_com:utils: Populate input parameters from config file

Sample config:

debug: True
csi: True
csi_format: "nexmon"
rssi: True
rssi_avg: True
rssi_poll_interval: 5
interface: "wlan0"
mac_addr_filter: "XX:XX:XX:XX:XX:XX"


* mesh_com:utils: Remove hardcoded parameters from the script

Populate channel/bandwidth from configuration file.


* mesh_com:utils: Add support for ESP CSI capture format

Add basic support for ESP CSI format over serial port.


* mesh_com:utils: Populate serial port from config file

Populate serial port configuration from config file for
ESP CSI capture mode.


* mesh_com:utils: Limit CSI record samples to max configurable count

Add support to limit ESP CSI record to max configurable count.


* mesh_com:utils: suffix timestamp in the CSI captured filename

Add timestamp in the captured CSI data file to avoid overwrite.


* mesh_parameters subscriber qos change (#67)

- cherry-pick to fix merge conflict from develop to main

* mesh_com: initial multi radio support script (#70)

* mesh_com: initial multi radio support script

* mesh_com: multi radio script help/usage updated

* mesh_com: frequency hardcoded for second radio

+++ modules/sc-mesh-secure-deployment/src/client-mesh.py
@@ -2,9 +2,11 @@
 
 import argparse
 import json
+import yaml
 import subprocess
 import time
 import hashlib
+import re
 
 import netifaces
 from getmac import get_mac_address
@@ -12,14 +14,30 @@ import requests
 from termcolor import colored
 from pathlib import Path
 
+
+# Get the mesh_com config
+print('> Loading yaml conf... ')
+try:
+    yaml_conf = yaml.safe_load(open('src/mesh_com.conf', 'r'))
+    conf = yaml_conf['client']
+    debug = yaml_conf['debug']
+    print(conf)
+except (IOError, yaml.YAMLError) as error:
+    print(error)
+    exit()
+
+
 # Construct the argument parser
 ap = argparse.ArgumentParser()
+
 
 # Add the arguments to the parser
 ap.add_argument("-s", "--server", required=True, help="Server IP:Port Address. Ex: 'http://192.168.15.14:5000'")
 ap.add_argument("-c", "--certificate", required=True)
 args = ap.parse_args()
 
+
+# Connect to server
 URL = args.server
 print('> Connecting to server: ' + str(URL))
 
@@ -33,7 +51,6 @@ def get_os():
             os = aux
     return os
 
-
 def get_data(cert_file, os):
     message = '/api/add_message/' + os
     response = requests.post(URL + message,
@@ -42,7 +59,8 @@ def get_data(cert_file, os):
         print(colored('Not Valid Certificate', 'red'))
         exit()
     else:
-        print('> Received encrypted message: ' + str(response.content))
+        if debug:
+            print('> Encrypted message: ' + str(response.content))
         with open('payload.enc', 'wb') as file:
             file.write(response.content)
 
@@ -51,10 +69,15 @@ def decrypt_response():  # assuming that data is on a file called payload.enc ge
     proc = subprocess.Popen(['src/ecies_decrypt', args.certificate], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     out, err = proc.communicate()
     aux_list = [element.decode() for element in out.split()]
+    # print(aux_list)
     server_cert = aux_list[:39]
-    new_list = aux_list[40:64]
+    # Get server configuration json using regex
+    new_list = ''.join(aux_list)
+    pattern = re.compile(r'\{(?:[^{}])*\}')
+    new_list = pattern.findall(new_list)
     output_dict = serializing(new_list)
-    print('> Decrypted Message: ', output_dict)  # res =  json.loads(output_dict)
+    if debug:
+        print('> Decrypted Message: ', output_dict)
 
     return output_dict, server_cert
 
@@ -86,62 +109,32 @@ def verify_certificate(old, new):
     return old_md5 == new_md5
 
 
-def create_config_openwrt(response):
-    res = json.loads(response)
-    if res['gateway']:
-        config_file = open('/etc/config/lime-node', 'a+')
+def get_interface(pattern):
+    interface_list = netifaces.interfaces()
+    interface = filter(lambda x: pattern in x, interface_list)
+    pre = list(interface)
+    if not pre:
+        print('> ERROR: Interface ' + pattern + ' not found!')
     else:
-        nodeId = int(res['addr'].split('.')[-1]) - 1  # the IP is sequential, then it gaves me the nodeId.
-        config_file = open('/etc/config/lime-node', 'w')
-        config_file.write('config lime ' + "'system'" + '\n')
-        config_file.write('\t' + 'option hostname ' + "'" + 'node' + str(nodeId) + "'" + '\n')
-        config_file.write('\t' + 'option firstbootwizard_configured ' + "'" + 'true' + "'" + '\n\n')
-    config_file.write('config wifi radio1' + '\n')
-    config_file.write('\t' + 'list modes ' + "'ieee80211s'" + '\n')
-    config_file.write('\t' + 'option ieee80211s_mesh_fwding ' + "'" + res['ieee80211s_mesh_fwding'] + "'" + '\n')
-    config_file.write('\t' + 'option ieee80211s_mesh_id ' + "'" + res['ieee80211s_mesh_id'] + "'" + '\n')
-    config_file.write('\t' + 'option channel ' + "'" + str(res['channel']) + "'" + '\n\n')
-    config_file.write('config lime ' + "'network'" + '\n')
-    config_file.write('\t' + 'list protocols ' + "'ieee80211s'" + '\n')
-    config_file.write('\t' + 'list protocols ' + "'lan'" + '\n')
-    config_file.write('\t' + 'list protocols ' + "'anygw'" + '\n')
-    config_file.write('\t' + 'list protocols ' + "'" + 'batadv:' + res['batadv'] + "'" + '\n')
-    config_file.write('\t' + 'list protocols ' + "'" + 'babeld:' + res['babeld'] + "'" '\n')
-    config_file.write('\t' + 'list protocols ' + "'" + 'bmx7:' + res['bmx7'] + "'" + '\n')
-    config_file.write('\t' + 'option main_ipv4_address ' + "'" + res['addr'] + '/24' + "'")
-    config_file.close()
-    subprocess.call('lime-config', shell=True)
-    time.sleep(2)
-    subprocess.call('reboot', shell=True)
-
-
-def get_ap_interface():
-    interface_list = netifaces.interfaces()
-    interface = filter(lambda x: 'wla' in x or 'wlp' in x, interface_list)
-    return list(interface)[0]
-
-
-def get_mesh_interface():
-    interface_list = netifaces.interfaces()
-    interface = filter(lambda x: 'wlx' in x, interface_list)
     if not list(interface):
         mesh_interface = filter(lambda x: 'wla' in x or 'wlp' in x, interface_list)
         return  list(mesh_interface)[0]
-    return list(interface)[0]
+        return pre[0]
 
 
-def ubuntu_gw(ap_inf):
-    print('> Configuring Ubuntu gateway node...')
+def ubuntu_gw(gw_inf):
+    print('> Configuring Ubuntu gateway node with gw_inf:' + str(gw_inf) + '...')
     # Create Gateway Service
     subprocess.call('sudo cp ../../common/scripts/mesh-gw.sh /usr/local/bin/.', shell=True)
     subprocess.call('sudo chmod 744 /usr/local/bin/mesh-gw.sh', shell=True)
     subprocess.call('sudo cp services/gw@.service /etc/systemd/system/.', shell=True)
     subprocess.call('sudo chmod 644 /etc/systemd/system/gw@.service', shell=True)
-    subprocess.call('sudo systemctl enable gw@' + str(ap_inf) + '.service', shell=True)
-    # Auto connect wlx to AP at boot using wpa_supplicant
-    subprocess.call('sudo cp conf/ap.conf /etc/wpa_supplicant/wpa_supplicant-' + str(ap_inf) + '.conf', shell=True)
-    subprocess.call('chmod 600 /etc/wpa_supplicant/wpa_supplicant-' + str(ap_inf) + '.conf', shell=True)
-    subprocess.call('sudo systemctl enable wpa_supplicant@' + str(ap_inf) + '.service', shell=True)
+    subprocess.call('sudo systemctl enable gw@' + str(gw_inf) + '.service', shell=True)
+    # IF inf is wl, auto connect wlx to AP at boot using wpa_supplicant
+    if "wl" in str(gw_inf):
+        subprocess.call('sudo cp conf/ap.conf /etc/wpa_supplicant/wpa_supplicant-' + str(gw_inf) + '.conf', shell=True)
+        subprocess.call('chmod 600 /etc/wpa_supplicant/wpa_supplicant-' + str(gw_inf) + '.conf', shell=True)
+        subprocess.call('sudo systemctl enable wpa_supplicant@' + str(gw_inf) + '.service', shell=True)
 
 
 def ubuntu_node(gateway):
@@ -168,42 +161,45 @@ def create_config_ubuntu(response):
         mesh_config.write('MAC=' + res['ap_mac'] + '\n')
         mesh_config.write('KEY=' + res['key'] + '\n')
         mesh_config.write('ESSID=' + res['ssid'] + '\n')
-        mesh_config.write('FREQ=' + res['frequency'] + '\n')
-        mesh_config.write('TXPOWER=' + res['tx_power'] + '\n')
+        mesh_config.write('FREQ=' + str(res['frequency']) + '\n')
+        mesh_config.write('TXPOWER=' + str(res['tx_power']) + '\n')
         mesh_config.write('COUNTRY=fi\n')
         mesh_config.write('PHY=phy1\n')
     # Are we a gateway node? If we are we need to set up the routes
-    if res['gateway']:
-        ap_interface = get_ap_interface()
-        ubuntu_gw(ap_interface)
-    else:
+    if res['gateway'] and conf['gw_service']:
+        print("============================================")
+        gw_inf = get_interface(conf['gw_inf'])
+        ubuntu_gw(gw_inf)
+    elif conf['dflt_service']:
         # We aren't a gateway node, set up the default route (to gw) service
         prefix = address.split('.')[:-1]
         prefix = '.'.join(prefix)
         mesh_gateway = prefix + '.2'
         ubuntu_node(mesh_gateway)
     # Set hostname
-    nodeId = int(res['addr'].split('.')[-1]) - 1  # the IP is sequential, then it gives the nodeId.
-    command_hostname = 'sudo hostnamectl set-hostname node' + str(nodeId)
-    subprocess.call(command_hostname, shell=True)
-    command_hostname_host = 'echo ' + '"' + address + '\t' + 'node' + str(nodeId) + '"' + ' >' + '/etc/hosts'
-    subprocess.call(command_hostname_host, shell=True)
+    if conf['set_hostname']:
+        print('> Setting hostname...')
+        nodeId = int(res['addr'].split('.')[-1]) - 1  # the IP is sequential, then it gives the nodeId.
+        subprocess.call('sudo hostnamectl set-hostname node' + str(nodeId), shell=True)
+        subprocess.call('echo ' + '"' + address + '\t' + 'node' + str(nodeId) + '"' + ' >' + '/etc/hosts', shell=True)
     # Final settings
-    subprocess.call('sudo nmcli networking off', shell=True)
-    subprocess.call('sudo systemctl stop network-manager.service', shell=True)
-    subprocess.call('sudo systemctl disable network-manager.service', shell=True)
-    subprocess.call('sudo systemctl disable wpa_supplicant.service', shell=True)
+    if conf['disable_networking']:
+        subprocess.call('sudo nmcli networking off', shell=True)
+        subprocess.call('sudo systemctl stop network-manager.service', shell=True)
+        subprocess.call('sudo systemctl disable network-manager.service', shell=True)
+        subprocess.call('sudo systemctl disable wpa_supplicant.service', shell=True)
     # Copy mesh service to /etc/systemd/system/
-    mesh_interface = get_mesh_interface()
-    subprocess.call('sudo cp ../../common/scripts/mesh-ibss.sh /usr/local/bin/.', shell=True)
-    subprocess.call('sudo chmod 744 /usr/local/bin/mesh-ibss.sh', shell=True)
-    subprocess.call('sudo cp services/mesh@.service /etc/systemd/system/.', shell=True)
-    subprocess.call('sudo chmod 664 /etc/systemd/system/mesh@.service', shell=True)
-    subprocess.call('sudo systemctl enable mesh@' + mesh_interface + '.service', shell=True)
-    # Ensure our nameserver persists as 8.8.8.8
-    subprocess.call('sudo cp conf/resolved.conf /etc/systemd/resolved.conf', shell=True)
-    time.sleep(2)
-    subprocess.call('reboot', shell=True)
+    if conf['mesh_service']:
+        mesh_interface = get_interface(conf['mesh_inf'])
+        subprocess.call('sudo cp ../../common/scripts/mesh-' + res['type'] + '.sh /usr/local/bin/.', shell=True)
+        subprocess.call('sudo chmod 744 /usr/local/bin/mesh-' + res['type'] + '.sh', shell=True)
+        subprocess.call('sudo cp services/mesh@.service /etc/systemd/system/.', shell=True)
+        subprocess.call('sudo chmod 664 /etc/systemd/system/mesh@.service', shell=True)
+        subprocess.call('sudo systemctl enable mesh@' + mesh_interface + '.service', shell=True)
+        # Ensure our nameserver persists as 8.8.8.8
+        # subprocess.call('sudo cp conf/resolved.conf /etc/systemd/resolved.conf', shell=True) # FIXME: I don't actually think we need this any more...
+        time.sleep(2)
+        subprocess.call('reboot', shell=True)
 
 
 if __name__ == "__main__":
@@ -213,12 +209,10 @@ if __name__ == "__main__":
     res, server_cert = decrypt_response()
     if verify_certificate(local_cert, server_cert):
         print(colored('> Valid Server Certificate', 'green'))
-        mac = get_mac_address(interface=get_mesh_interface())
+        mac = get_mac_address(interface=get_interface(conf['mesh_inf']))
         response = requests.post(URL + '/mac/' + mac)
         if os == 'Ubuntu':
             create_config_ubuntu(res)
-        if os == 'openwrt':
-            create_config_openwrt(res)
     else:
         print(colored("Not Valid Server Certificate", 'red'))
         exit(0)
